requirements.txt
pydantic==2.6.1
pytest==8.0.0
python-dotenv
